Unify example for Android and iOS
